Another large splitting, sim_reaction.c and simulation_workspace.c were created from simulation.c. Changed function signatures. Accuracy of new simulation routine needs to be evaluated still, some tests changed results slightly.
